Update burndown of Sprint 02
On the second Sprint day (after our SCRUM training) we updated our Sprint backlog. That
resulted in a lower number of story points (18). We also introduced a waterline and
added 2 cards there (10 storypoints).

The cards under the waterline were not properly show before. Same for the reduced number
of storypoints. This should be fixed now.