Remove current and old folders
Trollolo is implemented to have all the burndown charts in the same
folder. Due to have two folders we need to always generate the next
Sprint and then move the old Sprint to the old folder. This is not
efficient and also make the automation of the burndown chart generation
more complex.